+++ statham/dsl/elements/object.py
@@ -23,7 +23,6 @@ class Object(metaclass=ObjectMeta):
 
     poll = Poll({"questions": ["What's up?"]})
     ```
-    # TODO: patternProperties
     # TODO: propertyNames
     # TODO: minProperties
     # TODO: maxProperties