+++ tests/dsl/elements/test_properties.py
@@ -1,7 +1,7 @@
 import pytest
 
 from statham.dsl.elements import Element, Integer, Nothing, String
-from statham.dsl.elements.properties import Properties
+from statham.dsl.elements.properties import PatternDict, Properties
 from statham.dsl.property import _Property as Property
 
 
@@ -54,6 +54,23 @@ def test_properties_specific_additional():
     assert properties.additional == Integer()
 
 
+def test_properties_pattern():
+    parent = Element()
+    properties = Properties(
+        parent,
+        {"value": Property(String())},
+        {"^foo": Integer()},
+        additional=False,
+    )
+    assert "value" in properties
+    assert "foobar" in properties
+    assert "barfoo" not in properties
+    assert properties["value"].element == String()
+    assert properties["foobar"].element == Integer()
+    assert properties["foobar"].parent == parent
+    assert properties["foobar"].name == "foobar"
+
+
 @pytest.mark.parametrize(
     "additional,expected",
     [
@@ -64,13 +81,30 @@ def test_properties_specific_additional():
         (Integer(), ", additionalProperties=Integer()"),
     ],
 )
-def test_properties_repr(additional, expected):
+def test_properties_additional_repr(additional, expected):
     parent = Element()
     properties = Properties(
         parent, {"value": Property(String())}, additional=additional
     )
     assert repr(properties) == (
-        "Properties({'value': Property(String())}" f"{expected})"
+        "Properties({'value': Property(String())}" + f"{expected})"
+    )
+
+
+@pytest.mark.parametrize(
+    "pattern,expected",
+    [
+        ({}, ""),
+        ({"^foo": Integer()}, ", patternProperties={'^foo': Integer()}"),
+    ],
+)
+def test_properties_pattern_repr(pattern, expected):
+    parent = Element()
+    properties = Properties(
+        parent, {"value": Property(String())}, pattern=pattern
+    )
+    assert repr(properties) == (
+        "Properties({'value': Property(String())}" + f"{expected})"
     )
 
 
@@ -82,3 +116,35 @@ def test_properties_iter():
         False,
     )
     assert set(properties) == {"value", "other"}
+
+
+class TestPatternDict:
+    pattern_dict = PatternDict({"^foo": "bar"})
+
+    @staticmethod
+    @pytest.fixture(
+        params=[
+            ("foo", True),
+            ("foobar", True),
+            ("barfoo", False),
+            ("^foo", False),
+            (1, False),
+        ]
+    )
+    def param(request):
+        return request.param
+
+    def test_pattern_dict_contains(self, param):
+        key, valid = param
+        if valid:
+            assert key in self.pattern_dict
+        else:
+            assert key not in self.pattern_dict
+
+    def test_pattern_dict_getitem(self, param):
+        key, valid = param
+        if valid:
+            assert self.pattern_dict[key] == "bar"
+        else:
+            with pytest.raises(KeyError):
+                _ = self.pattern_dict[key]